Implement Discover actions and Rearrange actions for the bot.
Adding additional observation tensors for the discovery cards,
and using the plackett-luce distribution for the permutation of cards.

Also reworked the net internally to be responsible for the sampling
as well as the policy calculation, since now things are more complicated
and we don't just have a big conditional distribution. Now it is a big
stochastic neural net with internal sampling.