Create new copies of the bot only when the confidence interval doesn't contain the next best bot.
Previously we just compared the ratings directly, meaning we could create copies of a bot
if it was only epsilon better than the next best bot.

By making this change, we reduce the number of copies of bots we make when
they are not clearly better than previous versions, thus making it easier
to analyze bot performance over the long term.