first try at implementing the deathrattle queue
instead of resolving deathrattles in card.handle_event, the card is added to a deathrattle queue that is passed around via the combat phase context. Then the queue is emptied. Finally, we enter a loop that resolves deaths and deathrattles in sequence until there are no minions on the board that are in the process of dying.

This is pretty messy and definitely not complete - it doesn't work with death events not triggered though attacks, such as Red Whelp's start of combat trigger.

I'm also not sure if we want one deathrattle queue for the entire combat, or if we need to create new ones for each set of events that occur after an attack